Print error if output file is not writable
Since the default location for output files is /tmp, this can quickly occur
on systems shared by multiple users. At least print an error message so the
user knows that the files could not be properly generated.

While here, introduce a missing space into the error message for non-readable
skeletons.